Crate rattler_cache

Modules

package_cache

This module provides functionality to cache extracted Conda packages. See PackageCache.

run_exports_cache

This module provides functionality to cache extracted Conda packages. See RunExportsCache.

validation

Functionality to validate the contents of a Conda package.

Constants

PACKAGE_CACHE_DIR

The location in the main cache folder where the conda package cache is stored.

REPODATA_CACHE_DIR

The location in the main cache folder where the repodata cache is stored.

RUN_EXPORTS_CACHE_DIR

Functions

default_cache_dir

Determines the default cache directory for rattler. It first checks the environment variable RATTLER_CACHE_DIR. If not set, it falls back to the standard cache directory provided by dirs::cache_dir()/rattler/cache.

ensure_cache_dir

Creates the cache directory if it doesn’t exist and excludes it from backups.