rattler_cache

Module run_exports_cache

This module provides functionality to cache extracted Conda packages. See RunExportsCache.

Structs

BucketKey

A key that defines the actual location of the package in the cache.

CacheEntry

A cache entry that contains the path to the package and the run_exports.json

CacheKey

Provides a unique identifier for packages in the cache.

RunExportsCache

A RunExportsCache manages a cache of run_exports.json

Enums

CacheKeyError

RunExportsCacheError

An error that might be returned from one of the caching function of the RunExportsCache.