rattler_cache

Module validation

Functionality to validate the contents of a Conda package.

Almost all Conda packages contain a file info/paths.json that describes all the files the package contains. The validate_package_directory function validates that a directory containing an extracted Conda package archive actually contains the files as described by the paths.json file.

Very old Conda packages do not contain a paths.json file. These packages contain a (deprecated) files file as well as optionally a has_prefix and some other files. If the paths.json file is missing these deprecated files are used instead to reconstruct a [PathsJson] object. See [PathsJson::from_deprecated_package_directory] for more information.

Enums

PackageEntryValidationError

An error that indicates that a specific file in a package archive directory seems to be corrupted.

PackageValidationError

An error that is returned by validate_package_directory if the contents of the directory seems to be corrupted.

ValidationMode

The mode in which the validation should be performed.

Functions

validate_package_directory

Determine whether the files in the specified directory match what is expected according to the info/paths.json file in the same directory.

validate_package_directory_from_paths

Determine whether the files in the specified directory match wat is expected according to the passed in [PathsJson].