rattler_cache::validation

Function validate_package_directory

pub fn validate_package_directory(
    package_dir: &Path,
    mode: ValidationMode,
) -> Result<(IndexJson, PathsJson), PackageValidationError>

Determine whether the files in the specified directory match what is expected according to the info/paths.json file in the same directory.

If the info/paths.json file could not be found this function tries to reconstruct the information from older deprecated methods. See PathsJson::from_deprecated_package_directory.

If validation succeeds the parsed PathsJson object is returned which contains information about the files in the archive.